+++ php_fpm/tests/test_e2e.py
@@ -21,6 +21,7 @@ def test_status(dd_agent_check, instance, ping_url_tag):
         'php_fpm.requests.slow',
         'php_fpm.requests.accepted',
         'php_fpm.processes.max_reached',
+        'php_fpm.status.duration',
     ]
 
     expected_tags = ['fpm_cluster:forums', 'pool:www']
@@ -44,6 +45,7 @@ def test_status_fastcgi(dd_agent_check, instance_fastcgi, ping_url_tag_fastcgi):
         'php_fpm.requests.slow',
         'php_fpm.requests.accepted',
         'php_fpm.processes.max_reached',
+        'php_fpm.status.duration',
     ]
 
     expected_tags = ['fpm_cluster:forums', 'pool:www']