+++ datadog_checks_base/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 34.0.1 / 2023-10-17
+
 ***Fixed***:
 
 * Fix check cancellation timeout due to `DBMAsyncJob` cancellation being blocked ([#16028](https://github.com/DataDog/integrations-core/pull/16028))